requirements.txt
pandas>=2.0.0
yfinance>=0.2.18
PyPDF2>=3.0.0
pyxirr>=0.9.0
requests>=2.28.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
